changed lon long to int in loops

+++ Semester_2/Operating_Systems/2022/Assignment06/task2/task2.c
@@ -14,6 +14,7 @@ long long *consumer(void *cno)
 {
     long long *ret = malloc(sizeof(long long));
     long long sum = 0;
+    
     while (1)
     {
         pthread_mutex_lock(&mutex);
@@ -38,28 +39,28 @@ int main(void)
     pthread_t con[5];
     pthread_mutex_init(&mutex, NULL);
     void *status = 0;
-    long long name[5] = {1, 2, 3, 4, 5};
+    int name[5] = {1, 2, 3, 4, 5};
 
     myqueue_init(&queue);
 
-    for (long long i = 0; i < 5; ++i)
+    for (int i = 0; i < 5; ++i)
     {
         pthread_create(&con[i], NULL, (void *)consumer, (void *)&name[i]);
     }
 
-    for (long long i = 0; i < 10000; ++i)
+    for (int i = 0; i < 10000; ++i)
     {
         myqueue_push(&queue, 1);
     }
 
-    for (long long i = 0; i < 5; ++i)
+    for (int i = 0; i < 5; ++i)
     {
         myqueue_push(&queue, 0);
     }
 
     long long sum_total = 0;
     
-    for (long long i = 0; i < 5; ++i)
+    for (int i = 0; i < 5; ++i)
     {
         pthread_join(con[i], &status);
         sum_total += *(long long *)status;
